@@ -11,47 +11,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.12",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.12",
-    "@angular/compiler": "^7.2.12",
-    "@angular/core": "^7.2.12",
-    "@angular/forms": "^7.2.12",
-    "@angular/http": "^7.2.12",
-    "@angular/material": "^7.3.7",
-    "@angular/material-moment-adapter": "^7.3.7",
-    "@angular/platform-browser": "^7.2.12",
-    "@angular/platform-browser-dynamic": "^7.2.12",
-    "@angular/pwa": "^0.13.8",
-    "@angular/router": "^7.2.12",
-    "@angular/service-worker": "^7.2.12",
-    "core-js": "^2.6.5",
-    "g": "^2.0.1",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.0",
+    "@angular/material-moment-adapter": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/pwa": "^0.800.1",
+    "@angular/router": "^8.0.0",
+    "@angular/service-worker": "^8.0.0",
+    "core-js": "^3.1.3",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.2",
     "typescript": "^3.4.5",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.800.1",
     "@angular/cli": "^8.0.1",
-    "@angular/compiler-cli": "^7.2.12",
-    "@angular/language-service": "^7.2.12",
-    "@types/jasmine": "^2.8.16",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
+    "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^8.9.5",
+    "@types/node": "^12.0.4",
     "@types/web-bluetooth": "0.0.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "~5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.0.1",
+    "karma": "^4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.2",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1"
+    "ts-node": "~8.2.0",
+    "tslint": "~5.17.0"
   }
 }
