@@ -4,7 +4,7 @@ import { AvtoStruct } from '../_classes/avto';
 import { Subscription } from 'rxjs';
 import { GroupService, GroupStruct } from '../_services/groups.service';
 import { ServiceService, ServiceStruct } from '../_services/service.service';
-import { MatDialog, MatDialogConfig } from '@angular/material';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { finalize } from 'rxjs/operators';
 import { ScreenService, Screen, SmallScreen } from '../_services/screen.service';
 import { DialogCreateServiceComponent } from './dialogs/dialog-create-service/dialog-create-service.component';
