@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
-import { MatDialogRef, MatSnackBar } from '@angular/material';
+import { MatDialogRef } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { FormGroup, Validators, FormControl } from '@angular/forms';
 import { AvtoService } from 'src/app/_services/avto.service';
 import { AvtoStruct } from 'src/app/_classes/avto';
