@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { AvtoStruct } from '../../_classes/avto';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import { DialogUpdateAvtoComponent } from 'src/app/shared/dialog-update-avto/dialog-update-avto.component';
 
 @Component({
